Moves password code to passwords.py
