+++ users/arxiv/users/legacy/tests/test_util.py
@@ -1,13 +1,8 @@
 """Tests for :mod:`legacy_users.util`."""
 
 from unittest import TestCase
-from ..exceptions import PasswordAuthenticationFailed
 from .util import temporary_db
 from .. import util, models, sessions
-
-from hypothesis import given, settings
-from hypothesis import strategies as st
-import string
 
 class TestGetSession(TestCase):
     """
@@ -34,24 +29,3 @@ class TestGetSession(TestCase):
             self.assertEqual(tapir_session.session_id, int(session_id),
                              "Returned session has correct session id.")
 
-class TestCheckPassword(TestCase):
-    """
-    Tests passwords.
-    """
-    @given(st.text(alphabet=string.printable))
-    @settings(max_examples=500)
-    def test_check_passwords_successful(self, passw):
-        encrypted = util.hash_password(passw)
-        self.assertTrue( util.check_password(passw, encrypted.encode('ascii')),
-                         f"should work for password '{passw}'")
-
-    @given(st.text(alphabet=string.printable), st.text(alphabet=st.characters()))
-    @settings(max_examples=5000)
-    def test_check_passwords_fuzz(self, passw, fuzzpw):
-        if passw == fuzzpw:
-            self.assertTrue(util.check_password(fuzzpw,
-                                util.hash_password(passw).encode('ascii')))
-        else:
-            with self.assertRaises(PasswordAuthenticationFailed):
-                util.check_password(fuzzpw,
-                                    util.hash_password(passw).encode('ascii'))
